Test & bug fixes

+++ tests/co_occurrence/utils.py
@@ -44,13 +44,11 @@ def create_very_simple_tokens_pipeline(data: List[Tuple[str, List[str]]]) -> Cor
 
 
 def create_simple_bundle() -> Bundle:
-    tag: str = "TERRA"
     simple_corpus = very_simple_corpus(SIMPLE_CORPUS_ABCDEFG_3DOCS)
     context_opts: ContextOpts = ContextOpts(concept=set(), ignore_concept=False, context_width=2)
     bundle: Bundle = create_simple_bundle_by_pipeline(
         data=simple_corpus,
         context_opts=context_opts,
-        tag=tag,
     )
     return bundle
 
@@ -61,11 +59,15 @@ def create_simple_bundle_by_pipeline(
     tag: str = "TERRA",
     folder: str = None,
 ):
-    folder = os.path.join(folder or OUTPUT_FOLDER, tag)
-    shutil.rmtree(folder, ignore_errors=True)
+    folder = folder or OUTPUT_FOLDER
+    if folder.startswith('./tests') and folder != OUTPUT_FOLDER:
+        shutil.rmtree(folder, ignore_errors=True)
+
     if not isinstance(data, TokenizedCorpus):
         data: TokenizedCorpus = very_simple_corpus(data)
+
     config: CorpusConfig = CorpusConfig.tokenized_corpus_config()
+
     bundle: Bundle = (
         CorpusPipeline(config=config)
         .load_corpus(data)
